// union-talk-frontend/src/services/generated/file/models/assetFileRespVO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AssetFilePermissionRespVO } from './assetFilePermissionRespVO'

/**
 * 资产文件响应参数
 */
export interface AssetFileRespVO {
  /** 文件id */
  id?: string
  /** 会话id */
  conversationId?: string
  /** 文件夹id */
  folderId?: string
  /** 文件名称 */
  name?: string
  /** 文件扩展名 */
  fileExt?: string
  /** 文件类型 */
  fileType?: string
  /** 文件大小 */
  fileSize?: string
  /** MIME类型 */
  mimeType?: string
  /** 文件状态 */
  status?: string
  /** 创建时间 */
  createdAt?: string
  /** 当前用户文件权限 */
  permissions?: AssetFilePermissionRespVO
}