// union-talk-frontend/src/services/generated/file/models/assetFilePermissionRespVO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * 资产文件权限响应参数
 */
export interface AssetFilePermissionRespVO {
  /** 是否可以查看 */
  canView?: boolean
  /** 是否可以预览 */
  canPreview?: boolean
  /** 是否可以下载 */
  canDownload?: boolean
  /** 是否可以重命名 */
  canRename?: boolean
  /** 是否可以移动 */
  canMove?: boolean
  /** 是否可以删除 */
  canDelete?: boolean
}