// union-talk-frontend/src/services/generated/file/models/assetFolderPermissionRespVO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * 资产目录权限响应参数
 */
export interface AssetFolderPermissionRespVO {
  /** 是否可以查看 */
  canView?: boolean
  /** 是否可以上传文件 */
  canUpload?: boolean
  /** 是否可以创建子目录 */
  canCreateFolder?: boolean
  /** 是否可以重命名 */
  canRename?: boolean
  /** 是否可以移动 */
  canMove?: boolean
  /** 是否可以删除 */
  canDelete?: boolean
  /** 是否可以接收移动内容 */
  canAcceptMove?: boolean
}