// union-talk-frontend/src/services/generated/file/models/assetFolderTreeRespVO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AssetFolderPermissionRespVO } from './assetFolderPermissionRespVO'

/**
 * 资产目录树响应参数
 */
export interface AssetFolderTreeRespVO {
  /** 目录id */
  id?: string
  /** 会话id */
  conversationId?: string
  /** 父级目录id */
  parentId?: string
  /** 目录名称 */
  name?: string
  /** 目录层级 */
  levelNo?: number
  /** 目录类型 */
  folderType?: string
  /** 成员空间所有者用户id */
  ownerUserId?: string
  /** 当前用户目录权限 */
  permissions?: AssetFolderPermissionRespVO
  /** 子目录列表 */
  childFolderList?: unknown[]
}